Lighten shouldBehaveLike calling
